Use name-only path sensitivity for the javac task input
If you're trying to use the Gradle build cache across different
machines, then this input is currently problematic because any
difference in its absolute path will result in a different cache key.
The file it's checking is in your Gradle cache in your user home,
so it's probably not going to be the same across different people's
computers. Setting the path sensitivity to NAME_ONLY removes the
difference across environments while maintaining the check for
differing file contents.

Also add a property name to the input to assist with debugging
cache inputs; otherwise, it just shows up as "$1", which is not
terribly helpful.