Switch from Travis and AppVeyor to GitHub Actions
AppVeyor has been broken for a while anyway, so
this re-enables testing on Windows.

Build with JDK 8, run tests with JDK 11 and 14,
and with various Gradle versions.
On Windows, only test with JDK 11 with current Gradle.

This covers a smaller matrix than on Travis (where
each test Gradle version is only run with a single
JDK) but that should be enough to have good coverage.
Using JDK 8 as the baseline will trigger many more
tests (and more code paths), in all Gradle versions.