Update ErrorProne to 2.5.1
This is in preparation to testing with JDK 16.

The newer version of Caffeine brought in by the update
now has a "shadow" variant though, so the "manual" test
in non-java build fails as it cannot disambiguate them.
The test now applies the jvm-ecosystem plugin to setup
the variant matching rules, but as a result is now
limited to Gradle 6.7+ which introduced that plugin.
This really is testing an edge-case though, so it's OK.