Run Windows test with JDK 8 rather than JDK 11
The initial goal (IIRC) was to make sure that the build
could run with JDK 11, but unfortunately this means that
on Windows the JDK 8-specific -Xbootclasspath/p is not
execised, whereas this is probably the most important
thing to test for Windows compatibility, given that it
contains file paths.

With this change, the build itself (not the code) is no
longer exercised against JDK 11, but it doesn't really
matter actually.