Rework the way "JDK tests" are done
The integration test task is no longer run using a toolchain,
but toolchain is used to determine the installation path and
pass it as system property, that's then used in the TestKit
test project's gradle.properties as org.gradle.java.home.
The Java version is also passed as a system property to
replace JavaVersion.current() in tests.