Only assign options.isFork if needed
This is in preparation for Gradle 9 that will change isFork
to a lazy property (and dynamically rewrite the plugin
bytecode) which will make it immutable at execution time.

With this change, users can manually configure isFork=true
when needed to avoid the error, as the plugin will no longer
try to set isFork to true if it already is.

This is a first step towards fixing #97