Handle configurations more lazily
This is in preparation for Gradle 8.14 which will no longer
realize all configurations in the `base` plugin, allowing
truly lazily-initialized configurations.

This change should allow the javacConfiguration to only
be initialized when JDK 8 is being used, and the various
annotationProcessor configurations from source sets to
only be initialized when needed, e.g. when a compilation
tasks for that source set is run (it's likely that the
`java` plugin doesn't yet take advantage of this new
possibility itself though)